Move `note_and_explain_type_err` from `rustc_middle` to `rustc_infer`
This way we can properly deal with the types.
